kafka/app/requirements.txt
python-dotenv
numpy
scipy
matplotlib
scikit-learn
pandas
ipython 
jupyter-console
confluent-kafka
kafka-python
requests
beautifulsoup4
python-dotenv